adding tests for MongoV3 and Mongov4
