+++ src/test/java/com/github/vincentrussell/query/mongodb/sql/converter/AbstractQueryConverterIT.java
@@ -1,26 +1,17 @@
 package com.github.vincentrussell.query.mongodb.sql.converter;
 
+import com.github.vincentrussell.query.mongodb.sql.converter.rule.MongoRule;
 import com.google.common.base.Function;
 import com.google.common.base.Joiner;
 import com.google.common.base.Predicate;
 import com.google.common.collect.Collections2;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
-import com.mongodb.MongoClient;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.InsertOneModel;
 import com.mongodb.client.model.WriteModel;
-import de.flapdoodle.embed.mongo.MongodExecutable;
-import de.flapdoodle.embed.mongo.MongodProcess;
-import de.flapdoodle.embed.mongo.MongodStarter;
-import de.flapdoodle.embed.mongo.config.IMongodConfig;
-import de.flapdoodle.embed.mongo.config.MongoCmdOptionsBuilder;
-import de.flapdoodle.embed.mongo.config.MongodConfigBuilder;
-import de.flapdoodle.embed.mongo.config.Net;
-import de.flapdoodle.embed.mongo.distribution.Version;
 import org.apache.commons.io.IOUtils;
-import org.bson.BSON;
 import org.bson.BsonDocument;
 import org.bson.BsonString;
 import org.bson.Document;
@@ -28,54 +19,49 @@ import org.bson.json.JsonMode;
 import org.bson.json.JsonWriterSettings;
 import org.bson.types.ObjectId;
 import org.json.JSONException;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+import org.junit.Before;
 import org.junit.Test;
 import org.skyscreamer.jsonassert.JSONAssert;
 
 import java.io.BufferedReader;
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.StringWriter;
-import java.net.ServerSocket;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Random;
 
 import static org.junit.Assert.assertEquals;
 
-public class QueryConverterIT {
+public abstract class AbstractQueryConverterIT {
 
     private static final int TOTAL_TEST_RECORDS_PRIMER = 25359;
     private static final int TOTAL_TEST_RECORDS_CUSTOMERS = 599;
     private static final int TOTAL_TEST_RECORDS_FILMS = 1000;
     private static final int TOTAL_TEST_RECORDS_STORES = 2;
-    private static MongodStarter starter = MongodStarter.getDefaultInstance();
-    private static MongodProcess mongodProcess;
-    private static MongodExecutable mongodExecutable;
-    private static int port = getRandomFreePort();
-    private static MongoClient mongoClient;
+
     private static final String DATABASE = "local";
-    private static final String COLLECTION = "my_collection";
-    private static final String COLLECTION_CUSTOMERS = "my_collection_customers";
-    private static final String COLLECTION_FILMS = "my_collection_films";
-    private static final String COLLECTION_STORES = "my_collection_stores";
+    protected static final String COLLECTION = "my_collection";
+    protected static final String COLLECTION_CUSTOMERS = "my_collection_customers";
+    protected static final String COLLECTION_FILMS = "my_collection_films";
+    protected static final String COLLECTION_STORES = "my_collection_stores";
     private static final String DATASET_PRIMER = "primer-dataset";
     private static final String DATASET_FILMS = "films";
     private static final String DATASET_CUSTOMERS = "customers";
     private static final String DATASET_STORES = "stores";
-    private static MongoDatabase mongoDatabase;
-    private static MongoCollection mongoCollection;
-    private static JsonWriterSettings jsonWriterSettings = new JsonWriterSettings(JsonMode.STRICT, "\t", "\n");
+
+    protected static JsonWriterSettings jsonWriterSettings = JsonWriterSettings.builder().outputMode(JsonMode.RELAXED).indent(true).indentCharacters("\t").newLineCharacters("\n").build();
+    protected MongoDatabase mongoDatabase;
+    protected MongoCollection<Document> mongoCollection;
+
+    public abstract MongoRule getMongoRule();
 
     private static void loadRecords(int totalRecords, String dataset, MongoCollection mc) throws IOException {
     	List<Document> documents = new ArrayList<>(totalRecords);
-        try(InputStream inputStream = QueryConverterIT.class.getResourceAsStream("/" + dataset + ".json");
+        try(InputStream inputStream = AbstractQueryConverterIT.class.getResourceAsStream("/" + dataset + ".json");
             BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream))) {
 
             String line;
@@ -94,37 +80,23 @@ public class QueryConverterIT {
         }
     }
     
-    @BeforeClass
-    public static void beforeClass() throws IOException {
-        IMongodConfig mongodConfig = new MongodConfigBuilder()
-                .version(Version.Main.PRODUCTION)
-                .cmdOptions(new MongoCmdOptionsBuilder()
-            		.useNoPrealloc(false)
-            		.useSmallFiles(false)
-            		.build())
-                .net(new Net("localhost",port, false))
-                .build();
+    @Before
+    public void before() throws IOException {
 
-        mongodExecutable = starter.prepare(mongodConfig);
-        mongodProcess = mongodExecutable.start();
-        mongoClient = new MongoClient("localhost",port);
+        mongoDatabase = getMongoRule().getDatabase(DATABASE);
 
-
-        mongoDatabase = mongoClient.getDatabase(DATABASE);
-        
-        
         mongoCollection = mongoDatabase.getCollection(COLLECTION_CUSTOMERS);
 
         mongoCollection.deleteMany(new Document());
         loadRecords(TOTAL_TEST_RECORDS_PRIMER,DATASET_CUSTOMERS,mongoCollection);
         assertEquals(TOTAL_TEST_RECORDS_CUSTOMERS,mongoCollection.countDocuments());
-        
+
         mongoCollection = mongoDatabase.getCollection(COLLECTION_FILMS);
 
         mongoCollection.deleteMany(new Document());
         loadRecords(TOTAL_TEST_RECORDS_PRIMER,DATASET_FILMS,mongoCollection);
         assertEquals(TOTAL_TEST_RECORDS_FILMS,mongoCollection.countDocuments());
-        
+
         mongoCollection = mongoDatabase.getCollection(COLLECTION_STORES);
 
         mongoCollection.deleteMany(new Document());
@@ -139,12 +111,6 @@ public class QueryConverterIT {
         
     }
 
-    @AfterClass
-    public static void afterClass() {
-        mongoClient.close();
-        mongodProcess.stop();
-        mongodExecutable.stop();
-    }
 
     @Test
     public void likeQuery() throws ParseException, JSONException {
@@ -155,41 +121,49 @@ public class QueryConverterIT {
         Document firstDocument = documents.get(0);
         firstDocument.remove("_id");
         JSONAssert.assertEquals("{\n" +
-                "\t\"address\" : {\n" +
-                "\t\t\"building\" : \"351\",\n" +
-                "\t\t\"coord\" : [-73.98513559999999, 40.7676919],\n" +
-                "\t\t\"street\" : \"West   57 Street\",\n" +
-                "\t\t\"zipcode\" : \"10019\"\n" +
+                "\t\"address\": {\n" +
+                "\t\t\"building\": \"351\",\n" +
+                "\t\t\"coord\": [\n" +
+                "\t\t\t-73.98513559999999,\n" +
+                "\t\t\t40.7676919\n" +
+                "\t\t],\n" +
+                "\t\t\"street\": \"West   57 Street\",\n" +
+                "\t\t\"zipcode\": \"10019\"\n" +
                 "\t},\n" +
-                "\t\"borough\" : \"Manhattan\",\n" +
-                "\t\"cuisine\" : \"Irish\",\n" +
-                "\t\"grades\" : [{\n" +
-                "\t\t\t\"date\" : {\n" +
-                "\t\t\t\t\"$date\" : 1409961600000\n" +
+                "\t\"borough\": \"Manhattan\",\n" +
+                "\t\"cuisine\": \"Irish\",\n" +
+                "\t\"grades\": [\n" +
+                "\t\t{\n" +
+                "\t\t\t\"date\": {\n" +
+                "\t\t\t\t\"$date\": \"2014-09-06T00:00:00Z\"\n" +
                 "\t\t\t},\n" +
-                "\t\t\t\"grade\" : \"A\",\n" +
-                "\t\t\t\"score\" : 2\n" +
-                "\t\t}, {\n" +
-                "\t\t\t\"date\" : {\n" +
-                "\t\t\t\t\"$date\" : 1374451200000\n" +
+                "\t\t\t\"grade\": \"A\",\n" +
+                "\t\t\t\"score\": 2\n" +
+                "\t\t},\n" +
+                "\t\t{\n" +
+                "\t\t\t\"date\": {\n" +
+                "\t\t\t\t\"$date\": \"2013-07-22T00:00:00Z\"\n" +
                 "\t\t\t},\n" +
-                "\t\t\t\"grade\" : \"A\",\n" +
-                "\t\t\t\"score\" : 11\n" +
-                "\t\t}, {\n" +
-                "\t\t\t\"date\" : {\n" +
-                "\t\t\t\t\"$date\" : 1343692800000\n" +
+                "\t\t\t\"grade\": \"A\",\n" +
+                "\t\t\t\"score\": 11\n" +
+                "\t\t},\n" +
+                "\t\t{\n" +
+                "\t\t\t\"date\": {\n" +
+                "\t\t\t\t\"$date\": \"2012-07-31T00:00:00Z\"\n" +
                 "\t\t\t},\n" +
-                "\t\t\t\"grade\" : \"A\",\n" +
-                "\t\t\t\"score\" : 12\n" +
-                "\t\t}, {\n" +
-                "\t\t\t\"date\" : {\n" +
-                "\t\t\t\t\"$date\" : 1325116800000\n" +
+                "\t\t\t\"grade\": \"A\",\n" +
+                "\t\t\t\"score\": 12\n" +
+                "\t\t},\n" +
+                "\t\t{\n" +
+                "\t\t\t\"date\": {\n" +
+                "\t\t\t\t\"$date\": \"2011-12-29T00:00:00Z\"\n" +
                 "\t\t\t},\n" +
-                "\t\t\t\"grade\" : \"A\",\n" +
-                "\t\t\t\"score\" : 12\n" +
-                "\t\t}],\n" +
-                "\t\"name\" : \"Dj Reynolds Pub And Restaurant\",\n" +
-                "\t\"restaurant_id\" : \"30191841\"\n" +
+                "\t\t\t\"grade\": \"A\",\n" +
+                "\t\t\t\"score\": 12\n" +
+                "\t\t}\n" +
+                "\t],\n" +
+                "\t\"name\": \"Dj Reynolds Pub And Restaurant\",\n" +
+                "\t\"restaurant_id\": \"30191841\"\n" +
                 "}", firstDocument.toJson(jsonWriterSettings),false);
     }
 
@@ -228,42 +202,6 @@ public class QueryConverterIT {
         } finally {
             mongoCollection.deleteOne(new Document("_id", new ObjectId("54651022bffebc03098b4567")));
             mongoCollection.deleteOne(new Document("_id", new ObjectId("54651022bffebc03098b4568")));
-        }
-    }
-    
-    @Test
-    public void objectIdInQueryFnInside() throws ParseException, JSONException {
-        mongoCollection.insertOne(new Document("_id", new ObjectId("54651022bffebc03098b4567")).append("key", "value1"));
-        mongoCollection.insertOne(new Document("_id", new ObjectId("54651022bffebc03098b4568")).append("key", "value2"));
-        try {
-            QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select _id from " + COLLECTION
-                + " where _id IN (OID('54651022bffebc03098b4567'),OID('54651022bffebc03098b4568'))").build();
-            QueryResultIterator<Document> findIterable = queryConverter.run(mongoDatabase);
-            List<Document> documents = Lists.newArrayList(findIterable);
-            assertEquals(2, documents.size());
-            JSONAssert.assertEquals("{\n" + "\t\"_id\" : {\n" + "\t\t\"$oid\" : \"54651022bffebc03098b4567\"\n"
-                + "\t}\n" + "}", documents.get(0).toJson(jsonWriterSettings),false);
-            JSONAssert.assertEquals("{\n" + "\t\"_id\" : {\n" + "\t\t\"$oid\" : \"54651022bffebc03098b4568\"\n"
-                + "\t}\n" + "}", documents.get(1).toJson(jsonWriterSettings),false);
-        } finally {
-            mongoCollection.deleteOne(new Document("_id", new ObjectId("54651022bffebc03098b4567")));
-            mongoCollection.deleteOne(new Document("_id", new ObjectId("54651022bffebc03098b4568")));
-        }
-    }
-    
-    @Test
-    public void objectIdEqQueryFnInside() throws ParseException, JSONException {
-        mongoCollection.insertOne(new Document("_id", new ObjectId("54651022bffebc03098b4567")).append("key", "value1"));
-        try {
-            QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select _id from " + COLLECTION
-                + " where _id = OID('54651022bffebc03098b4567')").build();
-            QueryResultIterator<Document> findIterable = queryConverter.run(mongoDatabase);
-            List<Document> documents = Lists.newArrayList(findIterable);
-            assertEquals(1, documents.size());
-            JSONAssert.assertEquals("{\n" + "\t\"_id\" : {\n" + "\t\t\"$oid\" : \"54651022bffebc03098b4567\"\n"
-                + "\t}\n" + "}", documents.get(0).toJson(jsonWriterSettings),false);
-        } finally {
-            mongoCollection.deleteOne(new Document("_id", new ObjectId("54651022bffebc03098b4567")));
         }
     }
 
@@ -972,42 +910,6 @@ public class QueryConverterIT {
     }
     
     @Test
-    public void simpleInnerJoin() throws ParseException, JSONException, IOException {
-    	QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select t1.Phone as Phonet1, t2.managerStaffId as managerStaffIdt2 from "+COLLECTION_CUSTOMERS+" as t1 inner join " + COLLECTION_STORES + " as t2 on t1.Country = t2.Country").build();
-    	QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
-    	List<Document> results = Lists.newArrayList(distinctIterable);
-        assertEquals(5, results.size());
-        JSONAssert.assertEquals("[{\n" + 
-        		"	\"Phonet1\": \"247646995453\",\n" + 
-        		"	\"managerStaffIdt2\": \"1\"\n" + 
-        		"},{\n" + 
-        		"	\"Phonet1\": \"615964523510\",\n" + 
-        		"	\"managerStaffIdt2\": \"1\"\n" + 
-        		"},{\n" + 
-        		"	\"Phonet1\": \"145720452260\",\n" + 
-        		"	\"managerStaffIdt2\": \"1\"\n" + 
-        		"},{\n" + 
-        		"	\"Phonet1\": \"164414772677\",\n" + 
-        		"	\"managerStaffIdt2\": \"1\"\n" + 
-        		"},{\n" + 
-        		"	\"Phonet1\": \"856872225376\",\n" + 
-        		"	\"managerStaffIdt2\": \"1\"\n" + 
-        		"}]",toJson(results),false);
-    }
-    
-    @Test
-    public void simpleInnerJoinByTwoFields() throws ParseException, JSONException, IOException {
-    	QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select t1.Phone as Phonet1, t2.managerStaffId as managerStaffIdt2 from "+COLLECTION_CUSTOMERS+" as t1 inner join " + COLLECTION_STORES + " as t2 on t1.Country = t2.Country and t1.City = t2.City").build();
-    	QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
-    	List<Document> results = Lists.newArrayList(distinctIterable);
-        assertEquals(1, results.size());
-        JSONAssert.assertEquals("[{\n" + 
-        		"	\"Phonet1\": \"247646995453\",\n" + 
-        		"	\"managerStaffIdt2\": \"1\"\n" + 
-        		"}]",toJson(results),false);
-    }
-    
-    @Test
     public void simpleSubquery() throws ParseException, JSONException, IOException {
     	QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select * from(select borough, cuisine from "+COLLECTION+" limit 1)").build();
         QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
@@ -1150,268 +1052,7 @@ public class QueryConverterIT {
         		"	\"c\" : 5518\n" +
         		"}]",toJson(results),false);
     }
-    
-    
-    @Test
-    public void subqueryJoinByOneGetMaxOfGroup() throws ParseException, JSONException, IOException {
-    	QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select r.cuisine as cuisine, trest.totalrestaurats as total from "+COLLECTION+" as r inner join (select cuisine, count(*) as totalrestaurats from "+COLLECTION+" group by cuisine) as trest on r.cuisine = trest.cuisine order by trest.totalrestaurats asc, cuisine asc limit 15").build();
-    	QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
-        List<Document> results = Lists.newArrayList(distinctIterable);
-        assertEquals(15, results.size());
-        JSONAssert.assertEquals("[{\n" + 
-        		"	\"cuisine\": \"Californian\",\n" + 
-        		"	\"total\": 1\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Chilean\",\n" + 
-        		"	\"total\": 1\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Creole/Cajun\",\n" + 
-        		"	\"total\": 1\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Polynesian\",\n" + 
-        		"	\"total\": 1\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"CafÃ©/Coffee/Tea\",\n" + 
-        		"	\"total\": 2\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"CafÃ©/Coffee/Tea\",\n" + 
-        		"	\"total\": 2\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Iranian\",\n" + 
-        		"	\"total\": 2\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Iranian\",\n" + 
-        		"	\"total\": 2\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Hawaiian\",\n" + 
-        		"	\"total\": 3\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Hawaiian\",\n" + 
-        		"	\"total\": 3\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Hawaiian\",\n" + 
-        		"	\"total\": 3\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Soups\",\n" + 
-        		"	\"total\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Soups\",\n" + 
-        		"	\"total\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Soups\",\n" + 
-        		"	\"total\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Soups\",\n" + 
-        		"	\"total\": 4\n" + 
-        		"}]",toJson(results),false);
-    }
-    
-    
-    @Test
-    public void subqueryJoinByTwoGetMaxOfGroup() throws ParseException, JSONException, IOException {
-    	QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select r.cuisine as cuisine, r.borough as borough, brest.totalrestaurats as total from "+COLLECTION+" as r inner join (select cuisine, borough, count(*) as totalrestaurats from "+COLLECTION+" group by cuisine, borough) as brest on r.cuisine = brest.cuisine and r.borough = brest.borough order by r.cuisine asc, r.borough asc limit 15").build();
-    	QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
-        List<Document> results = Lists.newArrayList(distinctIterable);
-        assertEquals(15, results.size());
-        JSONAssert.assertEquals("[{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Brooklyn\",\n" + 
-        		"	\"total\": 1\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Manhattan\",\n" + 
-        		"	\"total\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Manhattan\",\n" + 
-        		"	\"total\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Manhattan\",\n" + 
-        		"	\"total\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Manhattan\",\n" + 
-        		"	\"total\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"African\",\n" + 
-        		"	\"borough\": \"Bronx\",\n" + 
-        		"	\"total\": 31\n" + 
-        		"}]",toJson(results),false);
-    }
-    
-    
-    @Test
-    public void twoSubqueriesJoinByOneAndTwoGetMaxOfTwoGroups() throws ParseException, JSONException, IOException {
-    	QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select r.cuisine as cuisine, r.borough as borough, trest.totalrestaurats as total, brest.totalrestaurats as local from "+COLLECTION+" as r inner join (select cuisine, count(*) as totalrestaurats from "+COLLECTION+" group by cuisine) as trest on r.cuisine = trest.cuisine inner join (select cuisine, borough, count(*) as totalrestaurats from "+COLLECTION+" group by cuisine, borough) as brest on r.cuisine = brest.cuisine and r.borough = brest.borough order by cuisine asc, r.borough asc limit 15").build();
-    	QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
-        List<Document> results = Lists.newArrayList(distinctIterable);
-        assertEquals(15, results.size());
-        JSONAssert.assertEquals("[{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Brooklyn\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 1\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Manhattan\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Manhattan\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Manhattan\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Manhattan\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 4\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"Afghan\",\n" + 
-        		"	\"borough\": \"Queens\",\n" + 
-        		"	\"total\": 14,\n" + 
-        		"	\"local\": 9\n" + 
-        		"},{\n" + 
-        		"	\"cuisine\": \"African\",\n" + 
-        		"	\"borough\": \"Bronx\",\n" + 
-        		"	\"total\": 68,\n" + 
-        		"	\"local\": 31\n" + 
-        		"}]",toJson(results),false);
-    }
-    
-    
-    @Test
-    public void joinInSubqueryByOne() throws ParseException, JSONException, IOException {
-    	QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select t.cuisine, max(t.total) as maxi from (select r.cuisine as cuisine, trest.totalrestaurats as total from "+COLLECTION+" as r inner join (select cuisine, count(*) as totalrestaurats from "+COLLECTION+" group by cuisine) as trest on r.cuisine = trest.cuisine order by trest.totalrestaurats desc, cuisine asc limit 15) as t group by t.cuisine").build();
-    	QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
-        List<Document> results = Lists.newArrayList(distinctIterable);
-        assertEquals(1, results.size());
-        JSONAssert.assertEquals("[{\n" + 
-        		"	\"cuisine\" : \"American \",\n" +
-        		"	\"maxi\" : 6183\n" +
-        		"}]",toJson(results),false);
-    }
-    
-    @Test
-    public void joinInSubqueryAndJoinAgain() throws ParseException, JSONException, IOException {
-    	QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select t.cuisine as cuisine, max(t.total) as maxi, count(*) as coxi from "+COLLECTION+" as r inner join (select r.cuisine as cuisine, r.borough as borough, trest.totalrestaurats as total from "+COLLECTION+" as r inner join (select cuisine, count(*) as totalrestaurats from "+COLLECTION+" group by cuisine) as trest on r.cuisine = trest.cuisine order by trest.totalrestaurats desc, cuisine asc, borough limit 15) as t on r.cuisine = t.cuisine and r.borough = t.borough group by t.cuisine").build();
-    	QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
-        List<Document> results = Lists.newArrayList(distinctIterable);
-        assertEquals(1, results.size());
-        JSONAssert.assertEquals("[{\n" + 
-        		"	\"maxi\": 6183,\n" + 
-        		"	\"coxi\": 6165,\n" + 
-        		"	\"cuisine\": \"American \"\n" + 
-        		"}]",toJson(results),false);
-    }
-    
-    @Test
-    public void countGroupByQueryHavingByCount() throws ParseException, IOException, JSONException {
-        QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select cuisine, count(cuisine) from "+COLLECTION+" WHERE borough = 'Manhattan' GROUP BY cuisine HAVING count(cuisine) > 500").build();
-        QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
-        queryConverter.write(System.out);
-        List<Document> results = Lists.newArrayList(distinctIterable);
-        assertEquals(4, results.size());
-        JSONAssert.assertEquals("[{\n" + 
-        		"	\"cuisine\": \"Chinese\",\n" + 
-        		"	\"count\": 510\n" +  
-        		"}," +
-        		"{\n" + 
-        		"	\"cuisine\": \"Italian\",\n" + 
-        		"	\"count\": 621\n" +  
-        		"}," +
-        		"{\n" + 
-        		"	\"cuisine\": \"Café/Coffee/Tea\",\n" + 
-        		"	\"count\": 680\n" +
-        		"}," +
-        		"{\n" + 
-        		"	\"cuisine\": \"American \",\n" + 
-        		"	\"count\": 3205\n" +
-        		"}]",toJson(results),false);
-    }
-    
+
     @Test
     public void countGroupByQueryHavingByCountAlias() throws ParseException, IOException, JSONException {
         QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select cuisine, count(cuisine) as cc from "+COLLECTION+" WHERE borough = 'Manhattan' GROUP BY cuisine HAVING cc > 500").build();
@@ -1434,38 +1075,6 @@ public class QueryConverterIT {
         		"{\n" + 
         		"	\"cuisine\": \"American \",\n" + 
         		"	\"cc\": 3205\n" +
-        		"}]",toJson(results),false);
-    }
-    
-    @Test
-    public void havingMultiCond() throws ParseException, JSONException, IOException {
-    	QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select cuisine, count(*) as c from "+COLLECTION+" where 1=1 group by Restaurant.cuisine having count(*) >= 5 and count(*) <= 6").build();
-    	QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
-        List<Document> results = Lists.newArrayList(distinctIterable);
-        assertEquals(2, results.size());
-        JSONAssert.assertEquals("[{\n" + 
-        		"	\"cuisine\": \"Nuts/Confectionary\",\n" + 
-        		"	\"c\": 6\n" +  
-        		"}," +
-        		"{\n" + 
-        		"	\"cuisine\": \"Czech\",\n" + 
-        		"	\"c\": 6\n" +  
-        		"}]",toJson(results),false);
-    }
-    
-    @Test
-    public void havingAliasMultiCond() throws ParseException, JSONException, IOException {
-    	QueryConverter queryConverter = new QueryConverter.Builder().sqlString("select cuisine as cuisine, count(*) as c from "+COLLECTION+" where 1=1 group by Restaurant.cuisine having c >= 5 and c <= 6").build();
-    	QueryResultIterator<Document> distinctIterable = queryConverter.run(mongoDatabase);
-        List<Document> results = Lists.newArrayList(distinctIterable);
-        assertEquals(2, results.size());
-        JSONAssert.assertEquals("[{\n" + 
-        		"	\"cuisine\": \"Nuts/Confectionary\",\n" + 
-        		"	\"c\": 6\n" +  
-        		"}," +
-        		"{\n" + 
-        		"	\"cuisine\": \"Czech\",\n" + 
-        		"	\"c\": 6\n" +  
         		"}]",toJson(results),false);
     }
     
@@ -1579,33 +1188,9 @@ public class QueryConverterIT {
         }
     }
 
-    private static int getRandomFreePort() {
-        Random r = new Random();
-        int count = 0;
 
-        while (count < 13) {
-            int port = r.nextInt((1 << 16) - 1024) + 1024;
 
-            ServerSocket so = null;
-            try {
-                so = new ServerSocket(port);
-                so.setReuseAddress(true);
-                return port;
-            } catch (IOException ioe) {
-
-            } finally {
-                if (so != null)
-                    try {
-                        so.close();
-                    } catch (IOException e) {}
-            }
-
-        }
-
-        throw new RuntimeException("Unable to find port");
-    }
-
-    private static String toJson(List<Document> documents) throws IOException {
+    protected static String toJson(List<Document> documents) throws IOException {
         StringWriter stringWriter = new StringWriter();
         IOUtils.write("[", stringWriter);
         IOUtils.write(Joiner.on(",").join(Lists.transform(documents, new com.google.common.base.Function<Document, String>() {
